mobilys-be/requirements.txt
Django==5.2.1
pandas==2.2.3
psycopg2-binary==2.9.10
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
setuptools==67.8.0
django-cors-headers
django-environ==0.10.0
requests==2.32.3
Shapely>=1.8.0
pyproj>=3.6.0
osmnx==2.0.5
geopy==2.4.1
alphashape==1.3.1
geojson==3.2.0
geopandas==1.1.1
pillow>=10.0.0
gunicorn>=21.2
whitenoise>=6.7
openpyxl